Create Net Interception
Will target local network such that if a user attempts to visit the target website freeaeskey.xyz, it will give a non-random aes key. Academic project to study network vulnerabilities, practice on intercepting encrypted network data without permission is illegal and will result in a maximum of 5 years in jail.